Add pathfinder as extension
