+++ Project.toml
@@ -10,14 +10,11 @@ ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 LogDensityProblems = "6fdf6af0-433a-55f7-b3ed-c6c6e0b8df7c"
 LogDensityProblemsAD = "996a588d-648d-4e1f-a8f0-a84b347e47b1"
 MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
-Optim = "429524aa-4258-5aef-a3af-852621145aeb"
 PEtab = "48d54b35-e43e-4a66-a5a1-dde6b987cf69"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-Pathfinder = "b1d3bc72-d0e7-4279-b92f-7fa5d6d2d454"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SimpleUnPack = "ce78b400-467f-4804-87d8-8f486da07d0a"
 StyledStrings = "f489334b-da3d-4c2e-b8f0-e476e12c162b"
@@ -25,10 +22,12 @@ StyledStrings = "f489334b-da3d-4c2e-b8f0-e476e12c162b"
 [weakdeps]
 AdaptiveMCMC = "717c3277-546c-407d-8270-09a39a0919a0"
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
+Pathfinder = "b1d3bc72-d0e7-4279-b92f-7fa5d6d2d454"
 
 [extensions]
 PEtabBayesAdaptiveMCMCExt = ["AdaptiveMCMC"]
 PEtabBayesAdvancedHMCExt = ["AdvancedHMC"]
+PEtabBayesPathfinderExt = ["Pathfinder"]
 
 [compat]
 AdaptiveMCMC = "0.1.4"
@@ -52,8 +51,8 @@ ModelingToolkitBase = "1.23.0"
 Optim = "2.0.1"
 OrdinaryDiffEqRosenbrock = "1.25.0"
 PEtab = "5"
-Printf = "1"
 Pathfinder = "0.9.31"
+Printf = "1"
 Random = "1.10.0"
 SimpleUnPack = "1.1.0"
 StyledStrings = "1"
@@ -67,11 +66,14 @@ Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
 IOCapture = "b5f81e59-6552-4d32-b1f0-c071b021bf89"
+LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 ModelingToolkitBase = "7771a370-6774-4173-bd38-47e70ca0b839"
+Optim = "429524aa-4258-5aef-a3af-852621145aeb"
 OrdinaryDiffEqRosenbrock = "43230ef6-c299-4910-a778-202eb28ce4ce"
+Pathfinder = "b1d3bc72-d0e7-4279-b92f-7fa5d6d2d454"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["AdaptiveMCMC", "AdvancedHMC", "Catalyst", "CSV", "DataFrames", "FiniteDifferences", "IOCapture", "OrdinaryDiffEqRosenbrock", "ModelingToolkitBase", "Random", "SafeTestsets", "Test"]
+test = ["AdaptiveMCMC", "AdvancedHMC", "Catalyst", "CSV", "DataFrames", "FiniteDifferences", "IOCapture", "LineSearches", "Optim", "OrdinaryDiffEqRosenbrock", "ModelingToolkitBase", "Pathfinder", "Random", "SafeTestsets", "Test"]
