+++ src/common.jl
@@ -8,9 +8,22 @@ function _to_petab_scale(
 
     for i in eachindex(x_petab_scale)
         priors_scale[i] === :parameter_scale && continue
-        x_petab_scale[i] = _transform_x(x_petab_scale[i], parameters_scale[i])
+        x_petab_scale[i] = PEtab.transform_x(
+            x_petab_scale[i], parameters_scale[i], to_xscale = true
+        )
     end
     return x_petab_scale
+end
+function _to_petab_scale(
+        draws::T, inference_info::PEtabBayes.InferenceInfo
+    )::T where T <: AbstractMatrix{<:Real}
+    draws_petab_scale = similar(draws)
+
+    for j in axes(draws, 2)
+        draws_petab_scale[:, j] .= _to_petab_scale(@view(draws[:, j]), inference_info)
+    end
+
+    return draws_petab_scale
 end
 
 """
@@ -33,7 +46,7 @@ function to_prior_scale(
             continue
         end
 
-        x_prior_scale[i] = _transform_x(x, parameters_scale[i]; reverse = true)
+        x_prior_scale[i] = PEtab.transform_x(x, parameters_scale[i]; to_xscale = false)
     end
     return x_prior_scale
 end
@@ -63,14 +76,15 @@ function _gradient_to_inference_scale!(
     return nothing
 end
 
+function _sample_prior(
+        rng::Random.AbstractRNG, x::T, log_target::PEtabBayes.PEtabBayesLogDensity,
+    )::T where T <: AbstractVector{<:Real}
+    petab_prob = log_target.prob
+    inference_info = log_target.inference_info
 
-function _transform_x(x::T, transform::Symbol; reverse::Bool = false)::T where {T <: Real}
-    if transform == :log10
-        return reverse ? exp10(x) : log10(x)
-    elseif transform == :log
-        return reverse ? exp(x) : log(x)
-    elseif transform == :log2
-        return reverse ? exp2(x) : log2(x)
-    end
+    new_guess = PEtab.get_startguesses(rng, petab_prob, 1)
+    _x = to_prior_scale(new_guess, log_target)
+
+    copyto!(x, inference_info.bijectors(_x))
     return x
 end
